♻️ Refactor: Status screen components
