+++ client/src/features/userPage/components/UserProfile.tsx
@@ -1,5 +1,5 @@
 import GradientBackground from '../../../components/GradientBackground/GradientBackground.tsx';
-import LoadingScreen from '../../../components/LoadingScreen/LoadingScreen.tsx';
+import LoadingScreen from '../../../ui/StatusScreens/LoadingScreen.tsx';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
 import { followUser, unfollowUser } from '../../auth/userThunks.ts';
 import ImageHeader from '../../../components/ImageHeader/ImageHeader.tsx';
@@ -8,8 +8,8 @@ import { useEffect } from 'react';
 import { getUser } from '../userPageThunks.ts';
 import styles from '../../../components/PlayHeader/PlayHeader.module.scss';
 import TransparentButton from '../../../components/PlayHeader/TransparentButton.tsx';
-import NotFound from '../../../components/ErrorScreens/NotFound.tsx';
-import ServerError from '../../../components/ErrorScreens/ServerError.tsx';
+import NotFound from '../../../ui/StatusScreens/NotFound.tsx';
+import ServerError from '../../../ui/StatusScreens/ServerError.tsx';
 
 const isFollowed = (id: string, followings: string[]) => {
   return followings.includes(id);
