+++ client/src/ui/StatusScreens/ServerError.tsx
@@ -1,8 +1,8 @@
 const ServerError = () => {
   return (
-    <div className="h-full flex flex-col items-center justify-center ">
+    <div className="h-full flex flex-col items-center justify-center">
       <h2 className="mb-2 text-4xl font-semibold">🔥 500</h2>
-      <p>Error: something went wrong! 🤫 Pls don't tell the dev!</p>
+      <p>Error: something went wrong!</p>
     </div>
   );
 };
